+++ test/tools/test.ml
@@ -4,4 +4,9 @@ let () =
   Tw_tools.Cascade_provenance.report ();
   Tw_tools.Tailwind_gen.with_stats @@ fun () ->
   Alcotest.run "tools"
-    [ Test_entrypoint.suite; Test_project.suite; Test_tailwind_gen.suite ]
+    [
+      Test_entrypoint.suite;
+      Test_project.suite;
+      Test_tailwind_gen.suite;
+      Test_parity_compare.suite;
+    ]